CORE: Refactor PVA-based caput
Currently, there are two different functions provided by the pva
interface in NICOS to write values to EPICS records: _put_pv and
_put_pv_blocking. However, these methods are (mostly) duplicates of each
other:

Calling _put_pv("PVname", 1, wait=True), results in mostly the same
behaviour as _put_pv_blocking("PVname", 1), with the only difference being
the timeout (which is set to a hardcoded default value in the case of
_put_pv_blocking and taken from the parameter epicstimeout in case of
_put_pv.

Except for missing option of specifying a custom timeout at callsite,
_put_pv can therefore do everything _put_pv_blocking does, but
alternatively also operate in a "fire and forget"-mode, where it does
not block the session (by setting wait = False).

This PR therefore removes _put_pv_blocking and replaces all calls to it
with equivalent calls to _put_pv. Additionally, the "wait" parameter is
removed and instead the option to specify the timeout is given. If a
timeout is given, wait is automatically True and if no timeout is given,
wait is automatically False. For both backends p4p and caproto, this
results in the same behaviour as before (when wait was explictely
given).

Additionally, a new method `_put_pv_readback_checked` has been added.
This function does not rely on the EPICS callback mechanism for
verification, but instead repeatedly reads out the PV and compares it
with the input value, blocking the session in the meantime. This is
useful if the record does not provide the callback mechanism (e.g. the
motor record).

Change-Id: I76aa418a74b0ed5a545a9b3f5b6f6d61739ac58e
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/37508
Reviewed-by: Stefan Mathis <stefan.mathis@psi.ch>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>