SINQ/camini: Read threshold while comparing with beam current
The neutra scientist expects to be able to use the threshold as a
pause functionality, and that should be able to change it during
a measurement. This is odd, but I see no harm in allowing that.

They scientist claim that this is how it worked in the past. How that
worked I don't know, it's very possible that a similar change existed
locally in the past but then got lost.

Change-Id: I0f9caffe07803faf86f2e6c5e69d30f7352cfd34
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/37997
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Stefan Mathis <stefan.mathis@psi.ch>
Reviewed-by: Edward Wall <edward.wall@psi.ch>
Reviewed-by: Alexander Söderqvist <alexander.soederqvist@psi.ch>