moke01/utils: update fitting function
There were several iterations on this.
The basic concept is that the ideal function would look like
two adjacent curves of shape `_/‾`.
The fitting should end with two horizontal lines for top and
bottom saturation areas.
The problem of achieving this algorithmically is that real
measured curves can have different type of distortions, and also
an instrument error noise, which makes very hard to filter
"relevant" measurement points from "non-relevant" ones.
The current fitting function works well on curves having small
instrument noise and a bit of inclination.
The current plan is to use this fitting function as a guess
which instrument scientist would be able to adjust through
some gui element on the plots.

Change-Id: I64ba87c0557a4c1d3d2da31c9f199f0523a24a9d
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/38104
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Konstantin Kholostov <k.kholostov@fz-juelich.de>